requirements_openrouter.txt
# ============================================================================
# OMA v3.0 - Dependências (OpenRouter Version)
# ============================================================================
#
# Versão simplificada usando OpenRouter API ao invés de modelos locais
# Total: ~200MB de downloads (vs 11GB da versão local)
#
# ============================================================================

# Core
python-dotenv==1.0.1
httpx==0.27.2
asyncio==3.4.3

# Multi-Agent Framework
langgraph==0.2.28
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.5

# LLM Utils
tiktoken==0.7.0
pydantic==2.9.2

# Media Processing
ffmpeg-python==0.2.0
Pillow==10.4.0

# TTS (Local - Coqui)
TTS==0.22.0
pydub==0.25.1

# Vector Store (para memory do supervisor)
chromadb==0.5.5

# Web Interface
gradio==4.44.1

# API (opcional)
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Utils
tqdm==4.66.5
python-multipart==0.0.12
requests==2.32.3

# ============================================================================
# INSTALAÇÃO
# ============================================================================
#
# pip install -r requirements_openrouter.txt
#
# ============================================================================
